refactor: move read_args logic to separate file
