+++ pypdfeditor/main.py
@@ -4,51 +4,7 @@ import argparse
 
 from type_definitions import Args, Command
 
-
-class Command(StrEnum):
-    SPLIT = "split"
-
-
-class SplitArgs(NamedTuple):
-    source_file: str
-    pages: str
-
-
-class Args(NamedTuple):
-    command: Command
-    options: SplitArgs
-
-
-def read_args() -> Args:
-    parser = argparse.ArgumentParser(description="Python PDF editor")
-    subparser: argparse._SubParsersAction = parser.add_subparsers(
-        dest="command",
-        required=True,
-    )
-
-    split_parser: argparse.ArgumentParser = subparser.add_parser(
-        Command.SPLIT, help="Split a PDF file into multiple files based on page ranges."
-    )
-    split_parser.add_argument("source_file", help="Path to the source PDF file")
-    split_parser.add_argument(
-        "-p",
-        "--pages",
-        help="Specify page ranges to split on (e.g. '1-5,8-10')",
-        required=True,
-    )
-
-    args: argparse.Namespace = parser.parse_args()
-
-    if args.command == Command.SPLIT:
-        return Args(
-            command=args.command,
-            options=SplitArgs(
-                source_file=args.source_file,
-                pages=args.pages,
-            ),
-        )
-    else:
-        raise argparse.ArgumentError(None, "Error: Invalid command")
+from read_args import read_args
 
 
 def main() -> None:
